move last operations
